requirements.txt
pyserial
threading